Provide sig bypass validation exploitation test
This simulates the H01 finding described in detail in [1]. Effectively
verifying we are not vulnerable to this exploit.

The test intercepts the SAML Response sent from the IdP and injects
malicious XML to the Signature element, exploiting the vulnerability.
The response is then sent to EngineBlock to process.

Depending on the presence of the safe xmlseclibs, EngineBlock would then
process this assertion without problem. Allowing the attacker to provide
an assertion of their liking. From xmlseclibs 3.0.4, the SAML2 library
is no longer vulnerable to this attack. And in this PR, EngineBlock will
also provide defence in depth protection against tampered with SAML
Responses.

See:
1. https://www.hackmanit.de/en/blog-en/82-xml-signature-validation-bypass-in-simplesamlphp-and-xmlseclibs
2. https://simplesamlphp.org/security/201911-01
3. https://nvd.nist.gov/vuln/detail/CVE-2019-3465