Provide SAMLResponse input validation
Several checks are performed on the SAMLResponse. This allows us to
reject clearly invalid responses before actually processing them. See
this as an early, high level validation process.

The tests give insignts in the actual validations that are performed in
the SamlresponseValidator.

This addresses finding H01 in the Security audit performed by Hackmanit