Verify received LoA
The received LoA MUST match or be higher than the required LoA. M01
showed us that EB does not enforce this important rule. Stepup Gateway
should have verified this. But we should verify this ourself!

The IdP SP combination is used to determine the required LoA. The
received LoA is then loaded and tested against the LoA requirement.

If the requirement is not met, an EngineBlock exception is raised. This
exception is not handled in a user friendly manner as this is probably a
situation where a malicious user is trying to manipulate the SAML
Response.

See: https://www.pivotaltracker.com/story/show/169392199