Separate IdP and EngineBlock decorators
Created a new decorator: ProxiedIdentityProvider. This decorator is used
in the IdPs metadata. As we want to display the specific IdP UIinfo,
which in the previous setup was overridden by the EB values.

The new decorator uses the EB config to be able to use the EB certs,
contact persons and sso location.

Tests where added and existing tests where modified.

The test_only_add_entity_id_hash_for_sso_service_url_if_not_eb_self was
removed from the EBIdentityProviderTest. This test is now available in
the ProxiedIdentityProviderTest.

https://www.pivotaltracker.com/story/show/169708158