Prevent double schema and hostname on logo
When the IdPs metadata is rendered, Manage configured IdP logo urls
are used. Which are URIs, in contrary to the configured EB logo which is
a relative path to the locally sourced logo.

To get both situations to render the logo uri correctly, the EBconfig
object was updated. It now uses the EB hostname to create a logo URI
the schema is set to HTTPS as we can assume TLS is used for Engine.

Finally the ui_info.xml.twig was updated to no longer prepend the schema
and hostname.

A IdPs Metadata.feature scenario is used to verify that the logos are
rendered without double hostnames.

https://www.pivotaltracker.com/story/show/169724838