Change levelIsHigherOrEqualTo signature
The parameter is changed to take a Loa instead of an integer
representation of a LoA level. This makes working with the method
slightly harder as you actually need to have a LoA instance to compare
it to. But on every occasion in the code we now have this was the case.