Fix stepup loa
When setting a Loa on a Saml2 Assertion the LOA VO
was set on the AuthnContextClassRef.  In certain
distro's this resulted in a false positive when
asserting the is_string method. Because the VO had
a __toString method.

The method is now removed and explicitly a string
is given as parameter while maintaining the VO in
Corto itself.

(cherry picked from commit 55aa1e4fd1e204575845805e1b63ff881cd6c68e)