Disassemble the EB debug feature
It was not used, and provided minimal additional information. Removing
it seemed logical.

Regarding the debug Twig extension, the var_dump function was removed.
But the print_r and var_export functions have been left in place. As the
debug route (used to test EB from the home page) still uses it to create
debug info for the helpdesk mail.

https://www.pivotaltracker.com/story/show/164757722