Assemble transparant_authn_context MFA entities
When the level of an MFA-entity is set to transparant_authn_context, a
TransparentMfaEntity is created instead of a regular MFA entity. This
can later be used to determine what behaviour should be expected from
EngineBlock.

A factory was introducedd that now builds the Mfa entities, the factory
methods from the entities themselves have been removed in favour of the
factory.