Enable skipped API tests
The previously skipped api tests failed on the inability to overwrite
the engineblock.features service. This is disallowed for obvious
reasons, by using a test stand in for the feature configuration this
problem can be resolved.

The test stand in has a setter to customize the contents of the
features. Note that this version of the configuration should only be
used in testing scenarios.

See: https://www.pivotaltracker.com/story/show/174085599