Add trusted proxy signing verification
When the trusted proxy does not sing its requests, we error out of the
single sing on flow. The error is generic and has a non descriptive
error message is raised.

A behat scenario that was added some time ago to test this behaviour,
but was adapted to verify a normal TP flow was removed in this commit.
In favour of the newly added scenario.

In addition to that, another existing scenario was updated to sing
requests, this was not required before this change. Enabling that test
to pass.

See https://www.pivotaltracker.com/story/show/153995599