Repair functional tests
Many changes where required to get the functional testing mocks /
fixtures and controllers to work with the SAML2 v4 changes.

Most notable are the way extensions are set. Previously we could simply
treat the Extension field as an array, performing direct array
operations on it. Now using the getter/setter requires the use of
array_merge to function correctly.