Send NoPassive status response back to issuing SP
These NoPassive status responses are sent by the IdP who is not
able/allowed to process a Passive authentication.

Previously we would show a custom error page. But on this occasion it's
better to let the SP decide how to proceed in this situation.

See: https://www.pivotaltracker.com/story/show/164238276