Address three more ->getIssuer() issues
Two of the missing getValue instances impacted logging. They did not
break the log entry, but would include a xml representation of the
issuer instead of only the EntityID.

The latter one was already fixed at the other end (AppSingleton) but
never putting an Issuer VO into the session to begin with allowed me to
remove the previously added check.

Thanks @pablothedude and @tvdijen for chiming in on this one!