Make key-id optional in IdP constructors
The rollover key was not present in the IdP metadat of EngineBlock. This
is taken care of in this commit.

An additional challenge arose: The Key Id is not always explicitly defined
when opening the IdP metadata, in fact it is only used in roll-over
scenarios. Setting the default by default caused the SSO URl to be
filled with key:default. Where this is not a requirement.

https://www.pivotaltracker.com/story/show/174672018