Add try/catch to finding of original SP
The original SP is not always returning an SP entity, but sometimes
throws an Entity not found exception. This is probably warranted, but
will be caught later in the process.

This was breaking for the debug authentication. Where the internal sp
entity id of EB is used as the issuer of the authn request. This yielded
an entity not found exception from the repositories as internal entities
are not stored in the database anymore.

Simply catching the exception and continuing with the issuer sp entity
id should solve this issue