Change test script to be a prepare script
It no longer starts the test run, as the preparation should take place
(on CI) on the phpfpm container. We can not both prepare the env + build
the theme AND run the tests. As that needs to run on the cypress
container.

So breaking away the test run, and renaming the script should make
things more intuitive. The test-integration code became slightly less
readable, but is still easy enough to understand