Add automatic a11y test and fix problems
Added a new command to get rid of the symfony debugbar during testing.

There was a contrast issue with the edit button, because the font wasn't large enough.  Increased the font size to resolve.

There was an issue with an extra open strong tag.  Removing it fixed that.

The search form lacked a submit button.  Fixed that while keeping the styling the same.