Put temporary code in place to make WAYF function
This change adds support for the Skeune WAYF design, which is WIP to
begin with. So the change to the wayf.js must be considered temporary
code.

In order to continue to the IdP, an entity Id must be set on the form.
This form consists of hidden form fields (request id, entityId and so
forth) the form is then submitted.

The other code is only accessible when using the old theme.