Add integration tests for the new WAYF (#910)
* Add automatic a11y test and fix problems

Added a new command to get rid of the symfony debugbar during testing.

There was a contrast issue with the edit button, because the font wasn't large enough.  Increased the font size to resolve.

There was an issue with an extra open strong tag.  Removing it fixed that.

The search form lacked a submit button.  Fixed that while keeping the styling the same.

* Improve focus styles for footer

Given the different background colour, the normal focus styles weren't sufficient.

* Add aria role to the search form

* Resolve the a11y issues of the idp

Add a type to the button
Add a reminder to prevent bubbling on the button, so the link doesn't fire upon pressing the button (won't work without JS, but for now ok).

* Add a cursor pointer to all buttons

* Add groundwork for the integration tests

* Add tests for all known wayf functionality

Skipping those that shouldn't work yet