Refactor JS for wayf (#918)
* Refactor JS for wayf page

Split up in more manageable chunks across different files.
Get rid of a bunch of unneccessary constants which only clutter up the place.
Move all openconext specific initialization to openconext.
Limit skeune code to what's been tested (for now).

* Make test less brittle

This worked when testing manually, but not when it ran in the pipeline.  Not sure what caused it as i couldn't find a difference.  But clicks without force:true in cypress are very brittle.

* Optimize working with classes by using classList

Faster & more readable

* Remove obsolete file

* Duplicate wayf module files for now

To make it less timeconsuming to refactor the code

* Refactor IdpFilter to remove class

As classes will only decrease performance & legibility

* Add hidden class & refactor noresults

* Show correct IdpList on load

* Rework idp to include the form

* Show correct elements based on nojs

* Add behaviour for add an account button

* Handle pressing the disable/delete button

Refactored a bit to make this easier

* Handle removing idp from previous selection

Currently without removing from the cookie as it's not yet been set.

* Time adding the mousehandlers correctly

Given that the previous selection of idps is initially hidden, we need to add the mousehandlers for that part only when it's shown.  Firefox can add handlers to hidden elements, but Chrome can't.