Show Poirot on page load
And hide the reset button visually. When typing, show the reset button
(remove the class).

The background has also been removed from the poirot element. This
prevents setting extra logic to color with the parent elements focus
style. E.g. when the search bar is focussed, you want the poirot to be
focussed too