Validate required fields (#975)
* Improve request access form behavior

Rudimentary JS validation logic was added to simulate the required form
field 'html5 feature'. Actions are added to hide/show the error messages
when a form field is invalid.

* Cover validation in tests

* Remove double behaviour handler

Buttons already trigger a click handler when pressing enter on them.  As such, the keyboard behaviour handler was unnecessary here.

* Fix focus styles of disabled IDP

* Fix failing buttons

Also generalized the be.visible and have.class and their opposite counterparts.

* Disable one failing test

Cost of fixing it was too expensive

Co-authored-by: Koen Cornelis <kcornelis@ibuildings.nl>