Fix Safari bug where hidden elements have height
When loading the WAYF with the previously selected section, there was a significant scroll below that section (with only the blue background).  Some two or three screens worth.

The bug was caused by Safari taking into account the height of some hidden elements to calculate the page height, despite them being hidden.  Two in particular were the cause:
- all children of the remaining section, hidden by use of the hidden attribute but with a height and a display.
- the submit button, hidden by the class visually-hidden and the attribute hidden.