Position error-container the right way
With the previous way, when the text was too long, the top part of the container was no longer visible.  This was because of the absolute positioning combined with the translate function.

The new way utilises flexbox to center the the error box, ensuring that the entire content is visible & centered.