Increase room for tooltipless attribute values (#1033)
The idea here is to allow enough room for at least a 40-character attribute value, both on all screen sizes. The reason being that we want to ensure that the identifier fits 95% of the time. Because identifiers tend to be persistent nameids, which are a sh1-hash, that means 40 characters.