Egalize not converted pages (#1060)
* Egalize index-page

* Egalize error-page

* Ensure icons are not italic

* Adjust docs

* Update selector in step definition

The number of links on the / of EB where no longer countable. By
changing the CSS selector to something less ambigious, this should be
fixed.

* Prevent reading undefined pageTitle variable

The / page does not carry the pageTitle into the twig template like
others do. This resulted in an error. Easily remedied by wrapping the
creation of the title with a 'does the pageTitle exist' condition.

Co-authored-by: Michiel Kodde <mkodde@ibuildings.nl>