Fix IE11 bugs with search
Prior to this change there were two problems in IE11:
- the search reset button did not appear
- the search/sort did not work

This change repairs the search & sort, as well as ensuring the reset button works as expected.  While working on the change it became apparent that the selectors of the test were not yet as easily replaceable as those from the js itself and so that was remedied while at it.

[Issue brought up as part of the feedback round](https://wiki.surfnet.nl/display/coininfra/Bevindingen+nav+tests+door+supportteam)