Couple cypress tests to same selectors (#1072)
Prior to this change cypress tests were entirely decoupled from the selectors used in the rest of the application.  This resulted in tests breaking when one of those selectors was changed.

This change imports the selectors from the same file as the one used by the application JS, so that both have a single source of truth.  In instances where special test selectors are needed, a test-selector file is used which composes it's selectors maximally by incorporating the "normal" selectors.

[Issue brought up as part of fixing the IE11 bug for the latest feedback round](https://wiki.surfnet.nl/display/coininfra/Bevindingen+nav+tests+door+supportteam)