Adjust styling of PEP page
Prior to this change the text in the error message was bolded.

This change changes the bold error message to a normal weight error message.  It also adds, in the readme, a link to the page containing functional testing endpoints for all different error pages, so they can be tested.
Last but not least an accessibility improvement was included: the logo now has an empty alt text.

[Issue discovered as part of fixing tickets for the latest feedback round](https://wiki.surfnet.nl/display/coininfra/Bevindingen+nav+tests+door+supportteam)