Ensure IE11 handles overflow in disclaimers
Prior to this change the overflow of text in the disclaimerlist was not ok for IE11 only.

This change fixes the overflow issue.  Turns out IE11 doesn't handle this well for grid children.  Setting a width on them resolves most of the issue.
For additional measure the word-wrap property was added for IE11 and the word-break property for other browsers.
Last, it was found that the code for disclaimer had a non-used double in components.  That double was removed.

[Issue discovered as part of fixing tickets for the latest feedback round](https://wiki.surfnet.nl/display/coininfra/Bevindingen+nav+tests+door+supportteam)