Fix IE11 bugs
Prior to this change, the connected idps in IE11 showed the disabled button even when not disabled.  This only happened on test2.

This change hides the delete-disabled button by default for remaining idps, and shows it only for those who are disabled.
It also cleans up a forgotten console.log statement, and adds some conditional logic to ensure that the focusOnArrowItem check only happens if there is a focused element (cleaning up the console in IE).

[Issue discovered as part of fixing tickets for the latest feedback round](https://wiki.surfnet.nl/display/coininfra/Bevindingen+nav+tests+door+supportteam)