Correct display of tooltip for group attributes
Prior to this change, goup attributes had their tooltip on every row.  The icon for the tooltip was below the value instead of next to it (as for normal attributes) and the tooltip become bold.

This change ensures the tooltip doesn't repeat, the icon is positioned properly and the text of the tooltip is normal weight.

Issue discovered as part of acceptation testing and [reported on the wiki](https://wiki.surfnet.nl/display/coininfra/Bevindingen+nav+tests+door+supportteam).