Keep WAYF operable when syncing cookies (#1081)
Prior to this change, it was possible that the WAYF was rendered inoperable when syncing cookies.  This happened when an idp was in the cookie, but not available when connecting to a service.  The resulting null error in the JS rendered the entire wayf inoperable.

This change ensures the cookie-syncing doesn't freeze the WAYF.  The problem should be solved with the extra null checks.  But as the sync should never freeze the wayf, it is now wrapped in a try/catch block to be on the safe side.

Issue discovered as part of acceptation testing.