Correct display of tooltip for group attributes (#1082)
* Correct display of tooltip for group attributes

Prior to this change, group attributes had their tooltip on every row.  The icon for the tooltip was below the value instead of next to it (as for normal attributes) and the tooltip become bold.

This change ensures the tooltip doesn't repeat, the icon is positioned properly and the text of the tooltip is normal weight.   It also adjust the test data so we can see if the group membership is rendered correctly.

Issue discovered as part of acceptation testing and [reported on the wiki](https://wiki.surfnet.nl/display/coininfra/Bevindingen+nav+tests+door+supportteam).

Co-authored-by: Michiel Kodde <mkodde@ibuildings.nl>