Ensure images after search are ok in IE11
Prior to this change, when searching all logos became the same placeholder logo in IE11 only.

This change ensures the image url is the correct one, so the logo is the correct one.

Issue discovered as part of acceptation testing and [reported on the wiki](https://wiki.surfnet.nl/display/coininfra/Bevindingen+nav+tests+door+supportteam).