Prevent logo flicker in FF when possible
Prior to this change, the logo flickers in FF whenever there's an interaction with the search bar.

This change limits the flicker to actual redraws of the idp list by preventing the search handler from being triggered if the previous search term matches the current one.

Issue discovered as part of acceptation testing and [reported on the wiki](https://wiki.surfnet.nl/display/coininfra/Bevindingen+nav+tests+door+supportteam).