Remove assertsversion tag from externally hosted images.
The IdP/Attribute source logo are hosted on a static vhost and do not
change as a result of a new EB software version. They may change for
other reasons but the EB assetversion has no relation to it. Remove the
assetsversion from the request so we do not needlessly invalidate these
logos when upgrading EB.