Log the destination when using unsolicited sso
The destination is not set on unsolicited sso authentications. This due
to the fact that it is not present on the original http request. This
data is loaded from the request uri (or redirect uri if the first is not
present). This data is only used for loggin purposes, but is valuable
for support desk personell as they can now see what IdP was used. Abeit it is
an md5 hashed string