Passthrough AuhtnFailed SamlResponses
When a SP request a AuthnContext and transparent_authn_context is
enabled, we pass on to the IDP. There is no way for us to know the
meaning or value for this AuthnContext. If a user can't use the
requested AuthnContext, the SP should be informed so it can decide what
to do.

https://www.pivotaltracker.com/story/show/173602383