Apply ARP only once in the filter chain.
The second time it was just done because ePTI is first added
unconditionally, then the entire ARP is applied to decide whether it is
let through.

The decision on whether to add the ePTI is added into the ePTI command
so we can avoid reapplying the entire ARP just for this goal.