Fix npm audit problems (#1142)
Fix most npm audit problems

There are quite a lot of security problems being thrown up by npm audit. This PR is to fix those.

After trying out a better fix, we opted to disable the security check for the devdependencies. Currently upgrading those is not compatible with maintaining support for IE8