Fix upgrade to v2
Prior to this change, there was a security problem with the flysystem
package.

This change upgrades the package from v1 to v2.  Given that this breaks
backwards compatibility, some changes needed to be made in the code to
account for this.

Original ticket made by dependabot: https://github.com/OpenConext/OpenConext-engineblock/pull/1149