Reset attributeValueTypes after an Attribute Manipulation
AttributeValueTypes contains the types (string, bool, int..) for each
value of all released attributes. The ARP ensures this remains
consistent after filtering out some values. However, the Attribute
Manipulation also has the option to change the released attributes in
the $attributes array passed to it. AM's do however not have a way to
change the corresponding list of types.

If this is not done, in the case that the resulting set has less values
than the original set after ARP, one gets an Exception from the SAML2
library that there are more valuetypes than values.

Earlier, this would always work because after AM the ARP would be
reapplied which kind of sorted out the attribute types. Not entirely
correctly btw, more by chance, but in practice this nearly always went
well because we never deal with anything but xsi:string. When the second
ARP step was removed, this reordering of the value types was also not
present anymore and the saml2 Exception appeared.

The solution is (and this makes it actually more correct than it was),
is that if anything changed about the attributes after AM, we just reset
the valuetypes. This triggers a feature in the saml2 library that if
there are no types, the library will assign the correct types based on
the PHP types of the values. This means that you can even decide in the
AM what types you'd want by setting $attributes correctly.