Move IsPassive check beyond the SSO notification and remember choice features.
A request with IsPassive requires no user interaction to take place. We check
if we have more than 1 IdP, if so, we need to present a WAYF (user interaction)
hence we throw a NoPassive response back in this case.

However, (since) two cases have been added below the check that will skip
the WAYF even if there's more than 1 choice: the SSO notification feature
and the remember choice feature. If any is true we continue directly to the IdP
so could support a passive request.

Therefore we move the IsPassive check to below these two cases. This means
we can support a few more scenarios in which an IsPassive request may
succeed.