Cover internal-CollabPersonId behaviour in test
The different scenarios with and without a trusted proxy are covered.

1. A normal situation where no TP is concerned should yield no internal
   collab person id
2. With TP, different NameId formats are not transformed, but remain
   unmangled in the assertion. The internal collab person id will
   contain an unspecified string representation of the users identifier
3. EPTI can still be present. It is set with the same value found in the
   NameId. So no trnasformations there either.

For now these changes have only been covered in behat tests. And a test
that was already touching this scenario was (re)moved from
SpProxy.feature.

Note that the behat test are not passing yet, at least the three ones
covering the new situation will not. They are marked WIP

See: https://www.pivotaltracker.com/story/show/181121279