Cover ARP logic in test when TP is involved
ARP tests did not factor in that a trusted proxy can potentially also
modify the set of attributes that are released. Adding the basic tests
found on top of the list of scenarios where copied and rewritten to also
check correct behavior when SP behind TP. And finally a test was added
that tests an intersection of attributes scoped by both TP and End-SP
are released.

See task 9 & 10 of https://www.pivotaltracker.com/story/show/177696789