In Consent, display NameID of the end-SP, not trusted proxy
When a trusted proxy is involved, serviceProviderMetadata contains
the information from the SP after the trusted proxy. This is used
for the SP name, ARP etc on the consent screen. Only the NameID
(format) was still read from the next-hop SP (the trusted proxy)
while the actually released nameId is the one configured for
the end-SP.