Prevent removal of form from idp when prev removed
Description of issue:
(Minor) If you press an unconnected IdP the RA form is shown, but you click the
greyed out IdP (which remains visible) again you get errors in the JS console

Solution:
Only remove the form from the IdP clone when the form still exists.
Clicking on a disconnected idp triggers the script to handle the request
access form interactions. One of these actions is to remove the form
triggering SSO on the targetted IdP. When the form is already removed in
a previous click, a successive click would trigger a console error.