Update JS dependencies
Quite some JS dependencies had not been updated for some time. Leaving
them vulnerable to security issues. Some of the packages have been
upgraded to a couple of major releases higher than currently installed.
This did not seem to have negative effect on the processes using the
packages. But please be aware of this.

For now only the socket.io and engine.io resolutions have been resolved.