Disable Cypress visual regression tests
After updating socket.io and engine.io, the snapshot plugin used by the
cypress snapshot plugin no longer works. In order to run the tests we
need the older (vulnerable) versions. And for passing security tests, we
need the new version.

(┛ಠ_ಠ)┛彡┻━┻

The current version of the snapshot project is not actively maintained,
but the community is trying to take it over from the original
maintainer. My suggestion is to keep it in the project for now, and hope
for a speedy upgrade. If this is not the case we could look to use
another plugin or disable the visual regression tests altogether?

In my endeavours of jugling cypress plugins, I also took the liberty to
update cypress to the latest version bumping it a whopping 4 major
releases. Another sign of caution is raised here.. Tests seemed to run
as before but we might experience BC breaking changes from the four
version upgrades.