Retrieve old/new style attribute hash in one go
Previously when testing if previous consent was given was based on
testing for old-style attribute hash calculation. And if that one did
not exist, the new (stable) attribute hash was tested.

Having that in two queries made little sense as it can easily be
performed in one action.

The interface of the repo and service changed slightly, as now a value
object is returned instead of a boolean value. The value object reflects
the version (type of attribute hash) of consent that was given. That can
either be: unstable, stable or not given at all.