Normalize attributes before validating them
This provides more accurate validation results, otherwise an IdP
sending only OID will fail all checks. The minor drawback is that
the debug page will show the already normalized names of the
attributes. This is not a big problem per se I think, because
in the end EB will work with this set, to what you send needs
to be normalizable to this.

Closes: #45