Support LoA 1.5
The config changed slightly to support loa 1.5. The array keys
indicating the numeric portion of a loa (1, 1.5, 2 and 3) used to be an
integer value. Well the index is now still integer, but multiplied by
10. This to support loa 1.5 (translate to 15 now)

Tests have been updated and extended with the new 1.5 loa.

The Loa and LoaRepository classes have been made more type-strict.