Clean up support for 'special' eb entities in service registry
In the past, the service regisry(manage) would contain special entities
for engineblock itself as an IdP and SP. This is no longer the case for
a long time now, the code related to these exceptions can be cleaned up.
The service registry only contains remote IdPs and SPs now.