Support ForceAuthn to Stepup Gateway callout
Manage can push SP Metadata containing the coin:stepup:forceauthn coin.

That value is picked up in the Assembler, and saved on the SP eb5 roles,
coin column.

During the StepUp round trip via StepUp Gateway, EB checks if the
FoceAuthn flag should be added to that AuthNRequest. If this is
specified for that SP, it is added. Otherwise no additional logic is
performed.

When dealing with a trusted proxy setup, the stepup settings for the
requesting sp are used, not that of the TP.

In order to get insghts in the outgoing AR to the StepUp Gateway, the
mock gateway page that is used in the funcitonal tests now displays the
received AR as an unsigned xml string. That string is used to perform
xpath queries on, determiniing the presense or absence of the forceauthn
flag.

See: https://www.pivotaltracker.com/story/show/184369636