Create Mdui UiInfo value objects
These objects represent the mdui prefixed metadata elements found in the
SP/IdP metadata. This data is multilingual. And is serialized in the new
mdui field that is to be introduced on the AbstractRole entity.

The current fields representing these values will be phased out in a
following release. For now we store those DisplayName, Description,
Logo,.. data on both the dedicated field and the new mdui field.
Allowing for a data-loss free roll back between this and the previous
version of EngineBlock.