Deprecate the Mdui fields that moved to Mdui field
The public fields will be removed in the next release. We need to keep
them for now to prevent data loss in a possible release rollback scenario.

And refrain from using these fields. This was quite the task as they are
written and read throughout the application and in tests.