When reporting an error, silence any session start notices
Because an error can happen at any point, a session may or may not have been started
so the code starts one in case it's not already there. It is not helpful / a red
herring to pollute error messages with "session already started" notices. Here
we just silence any notices since we're handling an error already, if we cannot
start the session for whatever reason we should not get distracted by this.