Show back to SP link on error page
When the preconditions are met, a Back to SP link is displayed on the
error feedback page. The preconditions are described and discussed on
the Pivotal story found in the link below.

The way this works is as follows: when EB sees that a response from the
IdP is not successful, it throws an exception. This results in a user
facing feedback page where the end user is instructed that
authentication failed for a given reason (they can vary (cancelled, no
auth context, ..). On that page the user is kind of stuck. But by
pressing the back to sp button, the user can now travel back to the SP
with the SamlFailedResponse. Allowing the SP to deal with the failed
authentication instead of remain stranded at Engine.

Some what was most challenging was how to port all required data to the
feedback page. You kind of need a lot of information to form a good Saml
failed response. And it turned out, passing along the full response from
the IdP was most convenient. This response is piggy backed to the
feedback controller via the FeedbackInfo session.

Not the most elegant of solutions..

What I did to make up for that is:
1. I added a test for the SamlResponseHelper
2. Produced some additional readme content, describing how to test
   EngineBlock step-up authentications

See: https://www.pivotaltracker.com/story/show/176960606