Specify data types of roles columns
The data type is based on the length you configure for the entity
fields. At least, that's the case for `object`, `array` and custom
types. By default a LONGTEXT is used to store that table data. When
running a MySQL like dbms in strict mode, that quickly becomes a problem
as LONGTEXT fields can contain many bytes of data. Adding the new MDUI
field was not permitted as the maximum table size restraint wat
triggered (IIRC).

To solve this issue, and to optimize the roles table. We decided to
review the data types for the text/object based fields. And set them to
a value that more realistically reflect their size.

As a result the following changes were made:

|**name**|**type**|**new length**|
|---|----|---|
|`certificates`|`array`|65535 (text)|
|`contact_persons`|`array`|65535 (text)|
|`name_id_formats`|`array`|65535 (text)|
|`single_sign_on_services`|`array`|65535 (text)|
|`shib_md_scopes`|`array`|65535 (text)|
|`attribute_release_policy`|`array`|65535 (text)|
|`assertion_consumer_services`|`array`|65535 (text)|
|`allowed_idp_entity_ids`|`array`|16777215 (mediumtext)|
|`requested_attributes`|`array`|65535 (text)|
|`organization_nl_name`|`object`|65535 (text)|
|`organization_en_name`|`object`|65535 (text)|
|`organization_pt_name`|`object`|65535 (text)|
|`consent_settings`|`json_array`|16777215 (mediumtext)|
|`consent_settings`|`object`|16777215 (text)|
|`mdui`|`engineblock_metadata_mdui`|65535 (text)|

Note that the logo field is not updated as it is marked for removal in
the next release.

See: https://www.pivotaltracker.com/n/projects/1425900/stories/185861244