Option to add ribbon to engineblock pages indication which environment you're working on
When left empty, nothing is shown.

The approach may look a bit unwieldy, but this is done so that envirnment
specific config can be passed to CSS. This is making it a pure CSS solution
that also would work in installs that have a CSP that disables inline styles,
that's why we have to enumerate the available colors since you cannot use
`attr()` with anything other than `content:`.