Request access: make form indicate that motivation is required
Currently upon submitting, the form is not accepted when motivation
is left empty, however there's no form feedback beforehand or in
fact anytime, the form just doesn't submit.

We fix this the easy way by adding the required '*' indicator and
using browser form validation to require the form to be filled in.