Improve social previews of EB URLs
This happens mostly when someone pastes a link to a social medium which
redirects immediately to OpenConext and e.g. shows a WAYF. The social
preview of that link is then the logo of the first IdP in the wayf. This
improves on that by ensuring that the logo used is the OpenConext instance's
own logo.

There are more fields available in the OpenGraph specification but we do
not have readily available content for those fields. The fields we add
now we can fill with information we already have in the template so are
a quick way to improve the social previews over the previous situation.